+++ Genius/Models/Settings/Defaults.Keys.swift
@@ -7,7 +7,7 @@ import Defaults
 
 extension Defaults.Keys {
 
-	static let showPassedMaintenanceChecks = Key<Bool>("ShowPassedMaintenanceChecks", default: false)
+	static let showResolvedRecommendations = Key<Bool>("ShowResolvedRecommendations", default: false)
 
 	static let useTextInsteadOfSymbols = Key<Bool>("UseTextInsteadOfSymbols", default: { Defaults[.interfaceMode] <= .simple })
 	static let interfaceMode = Key<Settings.InterfaceMode>("InterfaceMode", default: .normal)